cloudFunctions/instanceMetadata/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client
google-auth
google-auth-httplib2
oauth2client
google-cloud-storage 
google-cloud-bigquery